pyproject.toml
[project]
name = "datasearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "fastapi>=0.124.2",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "kaggle>=1.8.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]
